Commit 8
Implemented and tested SongDao with JUnit5
Changed IDs from Long to Integer

+++ SimFolk/src/sk/upjs/ics/mmizak/simfolk/core/dao/implementations/WeightedVectorDao.java
@@ -16,7 +16,7 @@ public class WeightedVectorDao implements IWeightedVectorDao {
     }
 
     @Override
-    public WeightedVector getWeightedVectorBySongId(Long songId, TermWeightType termWeightType,
+    public WeightedVector getWeightedVectorBySongId(Integer songId, TermWeightType termWeightType,
                                                     TermComparisonAlgorithm termComparisonAlgorithm, double tolerance) {
         return null;
     }
